@@ -1,11 +1,18 @@
 {
   "name": "@libresign/pdf-elements",
-  "description": "PDF viewer with draggable and resizable element overlays for Vue 2",
-  "version": "0.4.0",
+  "description": "PDF viewer with draggable and resizable element overlays for Vue 3",
+  "version": "1.0.0",
   "author": "LibreCode <contact@librecode.coop>",
   "private": false,
-  "main": "dist/pdf-elements.umd.js",
-  "module": "dist/pdf-elements.esm.js",
+  "main": "dist/index.js",
+  "module": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "types": "./dist/index.d.ts"
+    }
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/LibreSign/pdf-elements"
@@ -21,40 +28,58 @@
     "draggable",
     "resizable",
     "libresign",
-    "vue2"
+    "vue3"
   ],
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "build:lib": "vue-cli-service build --target lib --name pdf-elements src/index.js",
-    "lint": "vue-cli-service lint --no-fix",
-    "lint:fix": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "vite build",
+    "build:report": "vite build --mode report",
+    "build:demo": "vite build --mode demo",
+    "preview:demo": "vite preview --outDir dist-demo",
+    "test": "vitest run",
+    "test:watch": "vitest",
+    "lint": "eslint . --ext .vue,.ts,.js --max-warnings=0",
+    "lint:fix": "eslint . --ext .vue,.ts,.js --fix"
   },
   "dependencies": {
-    "pdfjs-dist": "^5.4.530",
-    "vue": "^2.7.16"
+    "pdfjs-dist": "^5.4.624",
+    "vue": "^3.5.28"
   },
   "devDependencies": {
-    "@babel/core": "^7.28.6",
-    "@babel/eslint-parser": "^7.28.6",
-    "@babel/plugin-transform-private-methods": "^7.28.6",
     "@nextcloud/browserslist-config": "^3.1.2",
-    "@vue/cli-plugin-babel": "^5.0.9",
-    "@vue/cli-plugin-eslint": "^5.0.9",
-    "@vue/cli-service": "^5.0.9",
-    "eslint": "^8.57.1",
-    "eslint-plugin-vue": "^9.24.0",
+    "@eslint/js": "^9.39.3",
+    "@typescript-eslint/eslint-plugin": "^8.56.0",
+    "@typescript-eslint/parser": "^8.56.0",
+    "@vitejs/plugin-vue": "^6.0.4",
+    "eslint": "^9.0.0",
+    "eslint-plugin-vue": "^10.8.0",
+    "eslint-plugin-vitest": "^0.5.4",
+    "globals": "^15.9.0",
+    "happy-dom": "^20.7.0",
     "postcss": "^8.5.6",
-    "vue-template-compiler": "^2.7.16"
+    "rollup-plugin-visualizer": "^7.0.0",
+    "typescript": "^5.9.3",
+    "vite": "^7.3.1",
+    "vite-plugin-dts": "^4.5.4",
+    "vitest": "^4.0.18",
+    "@vue/test-utils": "^2.4.6",
+    "vue-eslint-parser": "^10.4.0",
+    "vue-tsc": "^3.2.4"
   },
   "browserslist": [
     "extends @nextcloud/browserslist-config"
   ],
+  "overrides": {
+    "eslint-plugin-vitest": {
+      "@typescript-eslint/utils": "^8.56.0"
+    }
+  },
   "files": [
     "dist",
     "src",
     "COPYING",
     "README.md"
   ],
+  "sideEffects": false,
   "license": "AGPL-3.0-or-later"
 }
