@@ -11,11 +11,23 @@ module.exports = {
   extends: [
     'plugin:vue/essential',
     'eslint:recommended',
+    'plugin:@typescript-eslint/recommended',
   ],
   parserOptions: {
-    parser: '@babel/eslint-parser',
-    requireConfigFile: false,
+    parser: '@typescript-eslint/parser',
+    ecmaVersion: 2021,
+    sourceType: 'module',
   },
+  parser: 'vue-eslint-parser',
+  overrides: [
+    {
+      files: ['tests/**/*.ts'],
+      env: {
+        'vitest/globals': true,
+      },
+      plugins: ['vitest'],
+    },
+  ],
   rules: {
     'no-console': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
     'no-debugger': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
