@@ -6,10 +6,12 @@
   </button>
 </template>
 
-<script>
-export default {
+<script lang="ts">
+import { defineComponent } from 'vue'
+
+export default defineComponent({
   name: 'DeleteButton',
-}
+})
 </script>
 
 <style scoped>
