@@ -10,8 +10,10 @@
   </section>
 </template>
 
-<script>
-export default {
+<script lang="ts">
+import { defineComponent } from 'vue'
+
+export default defineComponent({
   name: 'DocumentsList',
   props: {
     documents: {
@@ -19,7 +21,7 @@ export default {
       default: () => [],
     },
   },
-}
+})
 </script>
 
 <style scoped>
