@@ -14,8 +14,10 @@
   </div>
 </template>
 
-<script>
-export default {
+<script lang="ts">
+import { defineComponent } from 'vue'
+
+export default defineComponent({
   name: 'SignatureBox',
   props: {
     object: {
@@ -27,7 +29,7 @@ export default {
       default: '14px',
     },
   },
-}
+})
 </script>
 
 <style scoped>
