@@ -1,7 +1,26 @@
 // SPDX-FileCopyrightText: 2026 LibreCode coop and contributors
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-export function getCachedMeasurement(cache, cacheKey, pageRef, pagesScale) {
+export interface CanvasMeasurement {
+  canvasWidth: number
+  canvasHeight: number
+}
+
+export interface PageMeasurement {
+  width: number
+  height: number
+}
+
+export interface PageMeasurementProvider {
+  getCanvasMeasurement: () => CanvasMeasurement
+}
+
+export function getCachedMeasurement(
+  cache: Record<string, PageMeasurement>,
+  cacheKey: string,
+  pageRef: PageMeasurementProvider,
+  pagesScale: number
+) {
   const cached = cache[cacheKey]
   if (cached) {
     return cached
