@@ -1,12 +1,14 @@
 // SPDX-FileCopyrightText: 2026 LibreCode coop and contributors
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-export function objectIdExistsInDoc(doc, objectId) {
+import type { PDFDocumentEntry, PDFElementObject } from '../types'
+
+export function objectIdExistsInDoc(doc: PDFDocumentEntry | undefined, objectId: string) {
   if (!doc || !objectId) return false
   return doc.allObjects.some((objects) => objects.some((obj) => obj.id === objectId))
 }
 
-export function findObjectPageIndex(doc, objectId) {
+export function findObjectPageIndex(doc: PDFDocumentEntry | undefined, objectId: string) {
   if (!doc || !objectId) return undefined
   for (let pageIndex = 0; pageIndex < doc.allObjects.length; pageIndex++) {
     if (doc.allObjects[pageIndex].some((obj) => obj.id === objectId)) {
@@ -16,7 +18,12 @@ export function findObjectPageIndex(doc, objectId) {
   return undefined
 }
 
-export function updateObjectInDoc(doc, pageIndex, objectId, payload) {
+export function updateObjectInDoc(
+  doc: PDFDocumentEntry | undefined,
+  pageIndex: number,
+  objectId: string,
+  payload: Partial<PDFElementObject>
+) {
   const objects = doc?.allObjects?.[pageIndex]
   if (!objects) return false
   const objectIndex = objects.findIndex((obj) => obj.id === objectId)
@@ -25,7 +32,11 @@ export function updateObjectInDoc(doc, pageIndex, objectId, payload) {
   return true
 }
 
-export function removeObjectFromDoc(doc, pageIndex, objectId) {
+export function removeObjectFromDoc(
+  doc: PDFDocumentEntry | undefined,
+  pageIndex: number,
+  objectId: string
+) {
   const objects = doc?.allObjects?.[pageIndex]
   if (!objects) return false
   const objectIndex = objects.findIndex((obj) => obj.id === objectId)
