@@ -1,13 +1,13 @@
 // SPDX-FileCopyrightText: 2026 LibreCode coop and contributors
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-export function getViewportWindow(scrollTop, viewHeight, margin = 300) {
+export function getViewportWindow(scrollTop: number, viewHeight: number, margin = 300) {
   return {
     minY: Math.max(0, scrollTop - margin),
     maxY: scrollTop + viewHeight + margin,
   }
 }
 
-export function isPageInViewport(offsetTop, offsetHeight, minY, maxY) {
+export function isPageInViewport(offsetTop: number, offsetHeight: number, minY: number, maxY: number) {
   return !(offsetTop + offsetHeight < minY || offsetTop > maxY)
 }
